requirements.txt
Django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.6
pillow==11.0.0